Removed hpe refs on stories (#4079)
